src/custom-metrics-example/requirements.txt
requests
opentelemetry-api
opentelemetry-sdk
prometheus_client
